op_walk: move "opset" query functions from jj_cli
